api/requirements.txt
Flask>=1.0.0
cryptography
python-jose[cryptography]
requests
soundfile
transformers
torch
faster-whisper==1.1.1
ctranslate2==4.6.0
ffmpeg-python>=0.2.0
google-cloud>=0.34.0
google-cloud-aiplatform>=1.0.0
